// web/src/api/generated/models/user.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CashLens
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  email?: string;
  is_active?: boolean;
  full_name?: string;
  avatar_url?: string;
  default_currency?: string;
  language_preference?: string;
  account_status?: string;
  id?: number;
  registration_date?: string;
  last_login_date?: string;
  two_factor_enabled?: boolean;
  two_factor_secret?: string;
}